switch to is app factory
- move APISession to ServiceSession
